+++ packages/common/document-models/index.ts
@@ -1,0 +1,1 @@
+export { module as DocumentDrive } from "./document-drive";