+++ packages/common/editors/generic-drive-explorer/generic-drive-explorer.stories.tsx
@@ -1,0 +1,9 @@
+import { createDriveStoryWithUINodes } from "editors/utils/storybook";
+import Editor from "./editor";
+
+const { meta, CreateDocumentStory: Empty } =
+  createDriveStoryWithUINodes(Editor);
+
+export { Empty };
+
+export default { ...meta, title: "Generic Drive Explorer" };