+++ packages/design-system/src/connect/components/breadcrumbs/breadcrumbs.stories.tsx
@@ -6,19 +6,19 @@ import {
   UiNodesContextProvider,
   useUiNodesContext,
 } from "@/connect";
+import { mockDriveNodes } from "@/connect/utils/mocks/ui-drive-node";
 import { Meta, StoryObj } from "@storybook/react";
 import { useEffect } from "react";
 import { Breadcrumbs } from ".";
-import { mockDriveNodes } from "@/connect/utils/mocks/ui-drive-node";
 
 const meta: Meta<typeof Breadcrumbs> = {
   title: "Connect/Components/Breadcrumbs",
   component: Breadcrumbs,
   decorators: [
-    (Story) => {
+    (Story, { args }) => {
       return (
         <UiNodesContextProvider>
-          <Story />
+          <Story {...args} />
         </UiNodesContextProvider>
       );
     },
@@ -27,14 +27,15 @@ const meta: Meta<typeof Breadcrumbs> = {
 
 export default meta;
 
-type Story = StoryObj<typeof meta>;
+type Story = StoryObj<typeof Breadcrumbs>;
 
 export const Default: Story = {
   args: {
-    isAllowedToCreateDocuments: true,
+    createEnabled: true,
   },
   render: function Wrapper(args) {
     const {
+      getNodeById,
       setDriveNodes,
       setSelectedNode,
       selectedNode,
@@ -82,13 +83,19 @@ export const Default: Story = {
       return Promise.resolve();
     }
 
+    const onCreateProp = args.createEnabled
+      ? {
+          onCreate: onAddAndSelectNewFolder,
+        }
+      : {};
+
     return (
       <div className="bg-white p-10">
         <Breadcrumbs
           {...args}
-          onAddAndSelectNewFolder={onAddAndSelectNewFolder}
-          selectedNodePath={selectedNodePath}
-          setSelectedNode={setSelectedNode}
+          breadcrumbs={selectedNodePath}
+          onBreadcrumbSelected={(b) => setSelectedNode(getNodeById(b.id))}
+          {...onCreateProp}
         />
       </div>
     );
@@ -99,6 +106,6 @@ export const NotAllowedToCreateDocuments: Story = {
   ...Default,
   args: {
     ...Default.args,
-    isAllowedToCreateDocuments: false,
+    createEnabled: false,
   },
 };