+++ packages/document-model-libs/editors/utils/storybook/document-story.tsx
@@ -1,21 +1,21 @@
 import { useArgs, useChannel } from "@storybook/preview-api";
-import { Meta, StoryObj } from "@storybook/react";
+import { Decorator, Meta, StoryObj } from "@storybook/react";
 import {
   Action,
   ActionContext,
   BaseAction,
+  Document,
   EditorProps,
   ExtendedState,
+  PartialState,
   Reducer,
   utils,
-  Document,
-  PartialState,
 } from "document-model/document";
 import React, { useState } from "react";
-import { useDocumentReducer } from "../reducer";
 import { useInterval } from "usehooks-ts";
+import { useDocumentReducer } from "../reducer";
 
-type EditorStoryArgs<S, A extends Action, L> = Partial<{
+export type EditorStoryArgs<S, A extends Action, L> = Partial<{
   isAllowedToCreateDocuments: boolean;
   isAllowedToEditDocuments: boolean;
   canUndo: boolean;
@@ -30,7 +30,11 @@ type EditorStoryArgs<S, A extends Action, L> = Partial<{
   };
 }>;
 
-type EditorStoryProps<S, A extends Action, L = unknown> = EditorProps<S, A, L> &
+export type EditorStoryProps<S, A extends Action, L = unknown> = EditorProps<
+  S,
+  A,
+  L
+> &
   EditorStoryArgs<S, A, L>;
 
 export type EditorStoryComponent<S, A extends Action, L = unknown> = (
@@ -46,6 +50,7 @@ export function createDocumentStory<S, A extends Action, L = unknown>(
   reducer: Reducer<S, A, L>,
   initialState: ExtendedState<PartialState<S>, PartialState<L>>,
   additionalStoryArgs?: EditorStoryArgs<S, A, L>,
+  decorators?: Decorator<EditorStoryProps<S, A, L>>[],
 ): { meta: Meta<typeof Editor>; CreateDocumentStory: DocumentStory<S, A, L> } {
   const meta = {
     component: Editor,
@@ -75,6 +80,7 @@ export function createDocumentStory<S, A extends Action, L = unknown>(
 
         return <Story emit={emit} />;
       },
+      ...(decorators ?? []),
     ],
     render: function Render(args) {
       const [error, setError] = useState<unknown>();